+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## TBD
+
+### Bug fixes
+
+* Ensure session and user information is included in native crash reports
+  [#138](https://github.com/bugsnag/bugsnag-unity/pull/138)
+  [bugsnag-cocoa#333](https://github.com/bugsnag/bugsnag-cocoa/pull/333)
+  [bugsnag-android#439](https://github.com/bugsnag/bugsnag-android/pull/439)
+
 ## 4.2.2 (2019-02-26)
 
 ### Bug fixes